add codebase conventions, extract command handlers and print/ingest runners from cli.tsx
